requirements.txt
pygame>=2.1.0
PyOpenGL>=3.1.5
numpy>=1.19.0 